// src/types.ts
export interface ApontamentoRow {
  "PARECER Nº": string;
  PARECERISTA: string;
  CONFERÊNCIA: string;
  URGÊNCIA: string;
  "TIPO DE PARECER": string;
  "TEMA DO PARECER": string;
  SUBTEMA: string;
  ESPÉCIE: string;
  "NATUREZA DO APONTAMENTO": string;
  APONTAMENTO: string;
  PROCEDIMENTO: string;
  "PARECER ANTERIOR": string;
  "LINK DO PARECER": string;
  "LINK REAL": string;
  EMENTA: string;
  JUSTIFICATIVA: string;
}